fix deleteUnusedConfigs

+++ cloud/linode/loadbalancers.go
@@ -212,13 +212,16 @@ func (l *loadbalancers) UpdateLoadBalancer(ctx context.Context, clusterName stri
 // Note: Don't build a map or other lookup structure here, it is not worth the overhead
 func (l *loadbalancers) deleteUnusedConfigs(ctx context.Context, nbConfigs []linodego.NodeBalancerConfig, servicePorts []v1.ServicePort) error {
 	for _, nbc := range nbConfigs {
+		found := false
 		for _, sp := range servicePorts {
 			if nbc.Port == int(sp.Port) {
-				continue
+				found = true
 			}
 		}
-		if err := l.client.DeleteNodeBalancerConfig(ctx, nbc.NodeBalancerID, nbc.ID); err != nil {
-			return err
+		if !found {
+			if err := l.client.DeleteNodeBalancerConfig(ctx, nbc.NodeBalancerID, nbc.ID); err != nil {
+				return err
+			}
 		}
 	}
 	return nil
